requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
slack_sdk==3.11.2
schedule==1.1.0
openai==0.10.2
slackeventsapi==3.0.0
